3接口完成

+++ src/components/manageUser.vue
@@ -140,15 +140,34 @@ export default {
       handleDelete(index, row) {
         console.log(index, row);
       },
-    exit: function () {
-      sessionStorage.clear()
-        this.$message({
+ exit: function () {
+      var that = this;
+             this.$axios.get('/logout', {
+
+  })
+  .then(function (response) {
+    console.log(response);
+      if (response.data.resultCode === 200) {
+        sessionStorage.clear()
+        that.$message({
             message: '退出成功',
             type: 'success',
             duration: 2000
           })
-          this.$router.push('/')
-    }
+          that.$router.push('/')
+       
+        } else {
+          that.$message({
+            message: '退出失败，可能是网络故障',
+            type: 'error',
+            duration: 2000
+          })
+        }
+  })
+  .catch(function (error) {
+    console.log(error);
+  });
+    } 
    
     },
     handleClose(tag) {
